# Speed up key fetching