--- a/Makefile
+++ b/Makefile
@@ -33,7 +33,7 @@ LDFLAGS =	-g $(MACHDEP) -Wl,-Map,$(notdir $@).map
 #---------------------------------------------------------------------------------
 # any extra libraries we wish to link with the project
 #---------------------------------------------------------------------------------
-LIBS	:=	 -lfat -lwiiuse -lbte -lm -logc -lruntimeiospatch
+LIBS	:=	 -lfat -lwiiuse -lbte -lm -logc
 
 #---------------------------------------------------------------------------------
 # list of directories containing libraries, this must be the top level containing