--- a/source/main.c
+++ b/source/main.c
@@ -1,7 +1,7 @@
+#include <ogc/machine/processor.h>
 #include <stdlib.h>
 #include <string.h>
 #include <gccore.h>
-#include <runtimeiospatch.h>
 
 #include "tools.h"
 
@@ -25,18 +25,13 @@ int main(int argc, char **argv)
     PrintHeadline();
 
     /* HW_AHBPROT check */
-    if (AHBPROT_DISABLED)
+    if (read32(HW_AHBPROT) == 0xFFFFFFFF)
     {
-        /* HW_AHBPROT flag is disabled */
-        printf("Applying runtime IOS patches, please wait...\n\n");
-        ret = IosPatch_RUNTIME(true, false, false, false);
-        if (ret > 0)
-        {
-            ret = XyzzyGetKeys();
-            if (ret != -2) printf("\nPress any button to exit.");
-        } else {
-            printf("Failed to apply runtime IOS patches! Press any button to exit.");
-        }
+        /* Make sure SRNPROT and MEM_PROT allow PPC access to SRAM and higher MEM2 */
+        mask32(HW_SRNPROT, 0, 0x8);
+        mask32(MEM_PROT, 0xFFFF0000, 0);
+        ret = XyzzyGetKeys();
+        if (ret != -2) printf("\nPress any button to exit.");
     } else {
         /* HW_AHBPROT flag is enabled */
         printf("The HW_AHBPROT hardware register is not disabled.\n");