… portlibs/include/runtimeiospatch.h
… portlibs/lib/libruntimeiospatch.a
… source/xxhash.h