* Removes runtimeiospatch - patching IOS isn't needed, and takes a few seconds upon app launch. Replaces with manually setting SRNPROT to allow PPC access and disabling memory protection via AHBPROT manually.
* Reads the SD key from the currently loaded ES module in MEM2, rather than loading individual IOS modules from NAND.
* When scanning RAM/files for the additional keys, uses XXHash rather than SHA-1. SHA-1 is still used to verify the key in case of any collision, but still speeds things up by around 3 seconds.

Some numbers for comparison (taken from vWii, IOS58 v6432, 5.5.5E - timer starts as soon as the first text displays, ends when keys display on-screen):
* xyzzy-mod 1.3.2 (release) - ~37.766s
* xyzzy-mod master - ~13.766s
* xyzzy-mod speedup - ~5.616s